(aws) allow customization of scaling policy copying
